requirements.txt
beautifulsoup4
pylucene